requirements.txt
Flask==3.0.3
gunicorn
google-cloud-firestore